requirements.txt
bitrix24-rest==1.1.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
convertdate==2.4.0
hijri-converter==2.2.4
holidays==0.13
idna==3.3
install==1.3.5
korean-lunar-calendar==0.2.1
PyMeeus==0.5.11
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
